Split widget: computed list
